[GITHUB-13] Allows ZK id to be set via tag
Adds ability to specify broker id via a tag upfront, this
removes the need for the aws cluster autodiscovery script
when using this method.

The idea is to use a single ASG per instance, the ASG specifies
the broker.id via a tag which is picked up the instance, in
addition an EBS volume and ENI keep the instances data and
IP address constant (hence the addition of the interface
options).

Other changes:

* Upgrade to verson 3.4.10 by default
* Makes tick options configurable, setting init time to
a bigger value gives more time for instances to elect a leader,
useful when starting a fresh cluster
* Removed sequential port assignment for hosts, this setting
is not needed
* Bumped to version 2 since the port changes may break
existing clusters